BILLING-337 pending not delete

+++ app/Http/Controllers/Api/InvoiceTableController.php
@@ -447,10 +447,25 @@ class InvoiceTableController extends Controller
         $item_meta_table = 'company_'.$request->company_id.'_item_metas';
         ItemMeta::setGlobalTable($item_meta_table);
         
+        // $status = InvoiceReceipt::where('invoice_id', $invoice->id)->where('paid', '0')->first();
+        // if($status){
+        //     return response()->json([
+        //         'status' => false,
+        //         'message' => "Can not delete unpaid invoice!"
+        //     ]);
+        // }
+
+        if($invoice->status == 'unpaid'){
+            return response()->json([
+                'status' => false,
+                'message' => "Can not delete unpaid invoice"
+            ]);
+        }
 
         Item::where('parent_id', $invoice->id)->where('type', $invoice->type)->delete();
         ItemMeta::where('parent_id', $invoice->id)->delete();
         InvoiceReceipt::where('invoice_id', $invoice->id)->delete();
+    
         if($invoice->delete()){
             return response()->json([
                 'status' => true,
